+++ src/atom-constructors/atom.ts
@@ -3,9 +3,7 @@ import cbShare from 'callbag-share'
 import cbMerge from 'callbag-merge'
 import cbTake from 'callbag-take'
 import { Atom, ReadableAtom, DerivedAtomReader, Updater, CustomAtom } from '..'
-import { atomToSource } from '../atom-to-source'
-import { equal } from '../inner-utils'
-import cachedSubject from '../cached-subject'
+import { equal, atomToSource, cachedSubject } from '../inner-utils'
 
 /**
  * The main atom constructor. Will return Atom, ReadableAtom or CustomAtom
