+++ src/inner-utils/equal-number-array.ts
@@ -1,0 +1,2 @@
+export const equalNumberArray = (l: number[], r: number[]) =>
+  l.length === r.length && !l.some((lVal, lIndex) => lVal !== r[lIndex])
