+++ src/inner-utils/identity.ts
@@ -1,0 +1,1 @@
+export const identity = (id: any) => id
