+++ src/inner-utils/index.ts
@@ -1,0 +1,11 @@
+/*
+These are implementation detail utilities, not meant to be exported from the Klyva package.
+*/
+
+export * from './get-atom-at-index'
+export * from './get-atom-at-key'
+export * from './equal'
+export * from './equal-number-array'
+export * from './identity'
+export * from './atom-to-source'
+export * from './cached-subject'
